Add unit tests for io collection keys tasks (#59)
* Remove unused test files

* Add testing dependencies

* Add docstring and unit tests for change key task

* Remove hello world

* Add docstring and unit tests for check key task

* Bump mypy (1.10.0 -> 1.11.0)

* Update test order to match source

* Add docstring and unit tests for copy key task

* Add docstring and unit tests for get keys task

* Bump setuptools (69.5.1 -> 72.1.0)

* Add docstring and unit tests for make key task

* Add docstring and unit tests for remove key task

+++ src/io_collection/keys/remove_key.py
@@ -4,6 +4,20 @@ import boto3
 
 
 def remove_key(location: str, key: str) -> None:
+    """
+    Removes object key at specified location.
+
+    Method will remove from the S3 bucket if the location begins with the
+    **s3://** protocol, otherwise it assumes the location is a local path.
+
+    Parameters
+    ----------
+    location
+        Object location (local path or S3 bucket).
+    key
+        Object key.
+    """
+
     if location[:5] == "s3://":
         remove_key_on_s3(location[5:], key)
     else:
@@ -11,10 +25,34 @@ def remove_key(location: str, key: str) -> None:
 
 
 def remove_key_on_fs(path: str, key: str) -> None:
+    """
+    Remove object key from local file system.
+
+    Parameters
+    ----------
+    path
+        Local object path.
+    key
+        Object key.
+    """
+
     full_path = os.path.join(path, key)
-    os.remove(full_path)
+
+    if os.path.isfile(full_path):
+        os.remove(full_path)
 
 
 def remove_key_on_s3(bucket: str, key: str) -> None:
+    """
+    Remove object key in AWS S3 bucket.
+
+    Parameters
+    ----------
+    path
+        AWS S3 bucket name.
+    key
+        Object key.
+    """
+
     s3_client = boto3.client("s3")
     s3_client.delete_object(Bucket=bucket, Key=key)
